refactor(client): move purchasing slice into features/purchasing
