+++ client/src/app/App.tsx
@@ -18,10 +18,10 @@ const Deliveries = lazy(() => import('../pages/Deliveries'));
 const SalesHistory = lazy(() => import('../features/sales/pages/SalesHistory'));
 const UsersPage = lazy(() => import('../features/admin/pages/Users'));
 const CustomersPage = lazy(() => import('../features/customers/pages/Customers'));
-const DistributorsPage = lazy(() => import('../pages/Distributors'));
+const DistributorsPage = lazy(() => import('../features/purchasing/pages/Distributors'));
 const CategoriesPage = lazy(() => import('../pages/Categories'));
 const SettingsPage = lazy(() => import('../features/admin/pages/Settings'));
-const PurchaseOrdersPage = lazy(() => import('../pages/PurchaseOrders'));
+const PurchaseOrdersPage = lazy(() => import('../features/purchasing/pages/PurchaseOrders'));
 const AuditLogPage = lazy(() => import('../features/admin/pages/AuditLog'));
 const PromotionsPage = lazy(() => import('../features/sales/pages/Promotions'));
 const GiftCardsPage = lazy(() => import('../features/sales/pages/GiftCards'));
@@ -30,7 +30,7 @@ const ExportsPage = lazy(() => import('../pages/Exports'));
 const CustomerDisplay = lazy(() => import('../pages/CustomerDisplay'));
 const RegisterPage = lazy(() => import('../pages/Register'));
 const ShiftsPage = lazy(() => import('../pages/Shifts'));
-const ExpensesPage = lazy(() => import('../pages/Expenses'));
+const ExpensesPage = lazy(() => import('../features/purchasing/pages/Expenses'));
 const SegmentsPage = lazy(() => import('../features/customers/pages/Segments'));
 const LayawayPage = lazy(() => import('../features/sales/pages/Layaway'));
 const BundlesPage = lazy(() => import('../pages/Bundles'));
@@ -40,7 +40,7 @@ const BranchesPage = lazy(() => import('../features/admin/pages/Branches'));
 const StorefrontPage = lazy(() => import('../pages/Storefront'));
 const OnlineOrdersPage = lazy(() => import('../pages/OnlineOrders'));
 const ReportBuilderPage = lazy(() => import('../pages/ReportBuilder'));
-const VendorsPage = lazy(() => import('../pages/Vendors'));
+const VendorsPage = lazy(() => import('../features/purchasing/pages/Vendors'));
 const SmartPricingPage = lazy(() => import('../pages/SmartPricing'));
 const AiInsightsPage = lazy(() => import('../pages/AiInsights'));
 const AdvancedAnalyticsPage = lazy(() => import('../pages/AdvancedAnalytics'));
